gc_worker: add integration tests to lock observer
