+++ tests/integrations/server/mod.rs
@@ -1,5 +1,6 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
+mod gc_worker;
 mod kv_service;
 mod lock_manager;
 mod raft_client;
